refactor: delegate falsify CI tester to the shared core (#278)
